staking comments added and types moved
